switched to rot2 instead of double, making a custom manifold for that purpose, after some point it fails with the ill-dependent variable error again, probably due to weird robot/banner angle coupling (increasing uncertainty on phi will probably help), associations still fucked

+++ all_seaing_perception/all_seaing_perception/factor_graph_slam.cpp
@@ -1379,7 +1379,7 @@ void FactorGraphSLAM::banners_cb(const all_seaing_interfaces::msg::LabeledObject
                                         std::sin(bearing + m_robot_pos_mean(2)),
                                         0);
             estimated_pose_vec[2] = all_seaing_perception::mod_2pi(m_robot_pos_mean(2) + phi);
-            gtsam::Pose2 new_banner_estimated_pose = gtsam::Pose2((gtsam::Vector)estimated_pose_vec.cast<double>());
+            gtsam::Pose2 new_banner_estimated_pose = gtsam::Pose2((gtsam::Vector)estimated_pose_vec);
             initialEstimate.insert(new_banner_key, new_banner_estimated_pose);
         }
         int tracked_id = match[i] >= 0 ? match[i] : m_num_banners - 1;
